refactor: share the live-process argv reader
The tunnel stop path already reads a pid's live command to prove identity
before signalling it. The collector paths need the same reader, so it moves
to process-args.ts with its tests; tunnel.ts keeps its own start token.
