+++ packages/stim-cli/src/engine/tunnel.ts
@@ -1,10 +1,18 @@
 import type { ChildProcess } from 'node:child_process';
-import { closeSync, mkdirSync, openSync, readFileSync, readSync, unlinkSync } from 'node:fs';
+import { closeSync, mkdirSync, openSync, readFileSync, unlinkSync } from 'node:fs';
 import { createHash, randomUUID } from 'node:crypto';
 import { basename, join, resolve as resolvePath, sep } from 'node:path';
 import { getConfigDir } from '../config.ts';
 import { getExecutor } from '../exec.ts';
 import { isPidAlive } from '../metro.ts';
+import {
+  PROCESS_COMMAND_MAX_BYTES,
+  PROCESS_COMMAND_TIMEOUT_MS,
+  type ReadProcCommand,
+  type RunPsCommand,
+  readProcFile,
+  readProcessArgs as readProcessArgsDefault,
+} from '../process-args.ts';
 import { createLineReader } from '../process-output.ts';
 import type { ManagedProvider } from './metro-reach.ts';
 import { withWorkspaceProcessLock, type WorkspaceProcessLockOptions } from './workspace-process-lock.ts';
@@ -565,9 +573,6 @@ export interface StopTunnelResult {
 
 const STOP_TIMEOUT_MS = 5_000;
 const STOP_POLL_MS = 100;
-const PROCESS_COMMAND_TIMEOUT_MS = 1_000;
-const PROCESS_COMMAND_MAX_BYTES = 32 * 1024;
-
 function defaultKill(pid: number): void {
   process.kill(pid, 'SIGTERM');
 }
@@ -576,115 +581,8 @@ function isEsrch(err: unknown): boolean {
   return (err as NodeJS.ErrnoException)?.code === 'ESRCH';
 }
 
-type ReadProcCommand = (path: string, maxBytes: number) => Buffer;
-type RunPsCommand = (pid: number, timeoutMs: number) => string;
-
-function defaultReadProcCommand(path: string, maxBytes: number): Buffer {
-  const fd = openSync(path, 'r');
-  const buffer = Buffer.alloc(maxBytes + 1);
-  let bytesRead = 0;
-  try {
-    while (bytesRead < buffer.length) {
-      const count = readSync(fd, buffer, bytesRead, buffer.length - bytesRead, null);
-      if (count === 0) break;
-      bytesRead += count;
-    }
-  } finally {
-    closeSync(fd);
-  }
-  if (bytesRead > maxBytes) throw new Error(`process command exceeds ${maxBytes} bytes`);
-  return buffer.subarray(0, bytesRead);
-}
-
-function defaultRunPsCommand(pid: number, timeoutMs: number): string {
-  return getExecutor().runFile('ps', ['-ww', '-o', 'command=', '-p', String(pid)], { timeoutMs });
-}
-
 function defaultRunPsStartCommand(pid: number, timeoutMs: number): string {
   return getExecutor().runFile('ps', ['-ww', '-o', 'lstart=', '-p', String(pid)], { timeoutMs });
-}
-
-function parseProcCommand(data: Buffer, maxBytes: number): string[] | null {
-  if (data.length === 0 || data.length > maxBytes || data[data.length - 1] !== 0) return null;
-  const args = data.subarray(0, -1).toString('utf-8').split('\0');
-  return args.length > 0 && args.every((arg) => arg.length > 0) ? args : null;
-}
-
-function parsePsCommand(command: string): string[] | null {
-  const input = command.trim();
-  if (!input || input.includes('\0') || input.includes('\n') || input.includes('\r')) return null;
-
-  const args: string[] = [];
-  let current = '';
-  let quote: "'" | '"' | null = null;
-  let escaped = false;
-  let started = false;
-  const finishArg = () => {
-    if (!started) return;
-    args.push(current);
-    current = '';
-    started = false;
-  };
-
-  for (const char of input) {
-    if (escaped) {
-      current += char;
-      started = true;
-      escaped = false;
-      continue;
-    }
-    if (char === '\\' && quote !== "'") {
-      escaped = true;
-      started = true;
-      continue;
-    }
-    if (quote) {
-      if (char === quote) quote = null;
-      else current += char;
-      started = true;
-      continue;
-    }
-    if (char === "'" || char === '"') {
-      quote = char;
-      started = true;
-      continue;
-    }
-    if (/\s/.test(char)) {
-      finishArg();
-      continue;
-    }
-    current += char;
-    started = true;
-  }
-  if (quote || escaped) return null;
-  finishArg();
-  return args.length > 0 && args.every((arg) => arg.length > 0) ? args : null;
-}
-
-export function readTunnelProcessArgs(
-  pid: number,
-  {
-    platform = process.platform,
-    readProcCommand = defaultReadProcCommand,
-    runPsCommand = defaultRunPsCommand,
-  }: {
-    platform?: NodeJS.Platform;
-    readProcCommand?: ReadProcCommand;
-    runPsCommand?: RunPsCommand;
-  } = {},
-): string[] | null {
-  try {
-    if (platform === 'linux') {
-      return parseProcCommand(
-        readProcCommand(`/proc/${pid}/cmdline`, PROCESS_COMMAND_MAX_BYTES),
-        PROCESS_COMMAND_MAX_BYTES,
-      );
-    }
-    if (platform === 'win32') return null;
-    return parsePsCommand(runPsCommand(pid, PROCESS_COMMAND_TIMEOUT_MS));
-  } catch {
-    return null;
-  }
 }
 
 function parseLinuxStartToken(data: Buffer, maxBytes: number): string | null {
@@ -708,7 +606,7 @@ export function readTunnelProcessToken(
   pid: number,
   {
     platform = process.platform,
-    readProcStat = defaultReadProcCommand,
+    readProcStat = readProcFile,
     runPsStartCommand = defaultRunPsStartCommand,
   }: {
     platform?: NodeJS.Platform;
@@ -763,7 +661,7 @@ export async function stopTunnel(
   record: TunnelRecord | null | undefined,
   {
     isAlive = isPidAlive,
-    readProcessArgs = readTunnelProcessArgs,
+    readProcessArgs = readProcessArgsDefault,
     readProcessToken = readTunnelProcessToken,
     kill = defaultKill,
     now = Date.now,
